+++ Cargo.toml
@@ -27,7 +27,7 @@ opt-level = 3
 bytemuck = "1.23.0"
 thiserror = "2.0.12"
 
-mesocarp = "0.8.4"
+mesocarp = "0.9.0"
 
 
 [dev-dependencies]
