+++ src/mt/agents.rs
@@ -26,7 +26,7 @@ pub struct PlanetContext<const INTER_SLOTS: usize, MessageType: Pod + Zeroable +
     /// all anti messages generated by this `Planet`
     pub anti_msgs: Journal,
     /// send counter for current time stamp
-    pub sends: Vec<Mail<MessageType>>,
+    pub(crate) sends: Vec<Mail<MessageType>>,
 }
 
 impl<const INTER_SLOTS: usize, MessageType: Pod + Zeroable + Clone>
