+++ src/mt/mod.rs
@@ -1,6 +1,6 @@
 //! Multi-threaded simulation execution with support for optimistic and conservative synchronization.
 //! Currently implements hybrid synchronization based on Clustered Time Warp architecture for
 //! parallel discrete event simulation across multiple threads.
-pub mod hybrid;
-pub mod optimism;
+
+pub mod fast;
 pub mod agents;
